Use deleteMany to clear all user documents in test
Previously, we were dropping the users collection before each test
which at the very least was bad for performance.

However the main reason for this change was that the disconnection
was not dropping the database properly, leaving a test database open
after the test suite was completed, potentially ruining the next
test run.