Move navigate to Signup and Login props
Using `useNavigate` inside the Signup and Login components was
causing component tests to fail since the `useNavigate` hook can
only be called within a `<BrowserRouter>` component.

We could have mounted each component within a `<BrowserRouter>`
comonent in the unit tests, but adding the navigate function as a
prop also allows for dependency injected mock functions to check
on the component level that the `navigate` function is called.