Add server diagnostic before and after e2e tests
My current hypothesis is the requests being sent at the start of
the e2e test are causing the server to shut down.
If the diagnostic after the e2e test fails, then the e2e tests are
causing the server to crash and some logging will be required to
check what the errors were.
  If this is the case we could catch uncaught errors in the server,
  output them to a file with fs, and then have another step after
  the e2e tests that `cat`s the file out the terminal for further
  diagnosis.
If the diagnostic after the e2e test succeeds, then the failing has
nothing to do with the server.