Add e2e tests for sending and receiving messages
NOTE: The second test is broken because the socket needs to be in
a ready state.
To ensure this happens, we're using a similar setInterval technique
to testing in the backend.
However it seems like the interval isn't being cleared properly
because cypress hangs after the test suite is run.
Either that or cypress isn't managing to access the chats page
properly.